Move TessellationOptions to options.py; minor renaming/reordering fixes.
